+++ src/main/java/no/finn/unleash/repository/FeatureToggleRepository.java
@@ -12,6 +12,7 @@ import no.finn.unleash.util.UnleashScheduledExecutor;
 
 public final class FeatureToggleRepository implements ToggleRepository {
     private final ToggleBackupHandler toggleBackupHandler;
+    private final ToggleBootstrapHandler toggleBootstrapHandler;
     private final ToggleFetcher toggleFetcher;
     private final EventDispatcher eventDispatcher;
 
@@ -29,7 +30,6 @@ public final class FeatureToggleRepository implements ToggleRepository {
                 toggleBackupHandler);
     }
 
-    @Deprecated
     public FeatureToggleRepository(
             UnleashConfig unleashConfig,
             UnleashScheduledExecutor executor,
@@ -39,8 +39,11 @@ public final class FeatureToggleRepository implements ToggleRepository {
         this.toggleBackupHandler = toggleBackupHandler;
         this.toggleFetcher = toggleFetcher;
         this.eventDispatcher = new EventDispatcher(unleashConfig);
-
-        toggleCollection = toggleBackupHandler.read();
+        this.toggleBootstrapHandler = new ToggleBootstrapHandler(unleashConfig);
+        this.toggleCollection = toggleBackupHandler.read();
+        if (this.toggleCollection.getFeatures().isEmpty()) {
+            this.toggleCollection = toggleBootstrapHandler.read();
+        }
 
         if (unleashConfig.isSynchronousFetchOnInitialisation()) {
             updateToggles().run();