+++ src/test/java/no/finn/unleash/repository/FeatureToggleRepositoryTest.java
@@ -1,12 +1,18 @@
 package no.finn.unleash.repository;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.*;
 
+import java.io.File;
+import java.io.IOException;
 import java.net.URISyntaxException;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import no.finn.unleash.ActivationStrategy;
 import no.finn.unleash.FeatureToggle;
@@ -42,6 +48,7 @@ public class FeatureToggleRepositoryTest {
                         .build();
 
         ToggleBackupHandler toggleBackupHandler = mock(ToggleBackupHandler.class);
+        when(toggleBackupHandler.read()).thenReturn(new ToggleCollection(Collections.emptyList()));
         ToggleFetcher toggleFetcher = mock(ToggleFetcher.class);
         UnleashScheduledExecutor executor = mock(UnleashScheduledExecutor.class);
         new FeatureToggleRepository(config, executor, toggleFetcher, toggleBackupHandler);
@@ -106,6 +113,8 @@ public class FeatureToggleRepositoryTest {
         ToggleFetcher toggleFetcher = mock(ToggleFetcher.class);
 
         ToggleBackupHandler toggleBackupHandler = mock(ToggleBackupHandler.class);
+        when(toggleBackupHandler.read()).thenReturn(new ToggleCollection(Collections.emptyList()));
+
         ToggleCollection toggleCollection =
                 populatedToggleCollection(
                         new FeatureToggle(
@@ -135,6 +144,7 @@ public class FeatureToggleRepositoryTest {
         UnleashScheduledExecutor executor = mock(UnleashScheduledExecutor.class);
         ToggleFetcher toggleFetcher = mock(ToggleFetcher.class);
         ToggleBackupHandler toggleBackupHandler = mock(ToggleBackupHandler.class);
+        when(toggleBackupHandler.read()).thenReturn(new ToggleCollection(Collections.emptyList()));
 
         // setup fetcher
         ToggleCollection toggleCollection = populatedToggleCollection();
@@ -158,6 +168,7 @@ public class FeatureToggleRepositoryTest {
         UnleashScheduledExecutor executor = mock(UnleashScheduledExecutor.class);
         ToggleFetcher toggleFetcher = mock(ToggleFetcher.class);
         ToggleBackupHandler toggleBackupHandler = mock(ToggleBackupHandler.class);
+        when(toggleBackupHandler.read()).thenReturn(new ToggleCollection(Collections.emptyList()));
 
         // setup fetcher
         ToggleCollection toggleCollection = populatedToggleCollection();
@@ -174,5 +185,67 @@ public class FeatureToggleRepositoryTest {
         List<FeatureToggle> list = new ArrayList();
         list.addAll(Arrays.asList(featureToggles));
         return new ToggleCollection(list);
+    }
+
+    @Test
+    public void should_read_from_bootstrap_location_if_backup_was_empty()
+            throws URISyntaxException, IOException {
+        File file =
+                new File(getClass().getClassLoader().getResource("unleash-repo-v1.json").toURI());
+        ToggleBootstrapProvider toggleBootstrapProvider = mock(ToggleBootstrapProvider.class);
+        when(toggleBootstrapProvider.read()).thenReturn(fileToString(file));
+        UnleashConfig config =
+                UnleashConfig.builder()
+                        .synchronousFetchOnInitialisation(false)
+                        .appName("test-sync-update")
+                        .unleashAPI("http://localhost:8080")
+                        .toggleBootstrapProvider(toggleBootstrapProvider)
+                        .build();
+        UnleashScheduledExecutor executor = mock(UnleashScheduledExecutor.class);
+        ToggleFetcher toggleFetcher = mock(ToggleFetcher.class);
+        ToggleBackupHandler toggleBackupHandler = mock(ToggleBackupHandler.class);
+        when(toggleBackupHandler.read()).thenReturn(new ToggleCollection(Collections.emptyList()));
+        FeatureToggleRepository repo =
+                new FeatureToggleRepository(config, executor, toggleFetcher, toggleBackupHandler);
+        assertThat(repo.getFeatureNames()).hasSize(3);
+    }
+
+    @Test
+    public void should_not_read_bootstrap_if_backup_was_found()
+            throws IOException, URISyntaxException {
+        File file =
+                new File(getClass().getClassLoader().getResource("unleash-repo-v1.json").toURI());
+        ToggleBootstrapProvider toggleBootstrapProvider = mock(ToggleBootstrapProvider.class);
+        when(toggleBootstrapProvider.read()).thenReturn(fileToString(file));
+        UnleashConfig config =
+                UnleashConfig.builder()
+                        .synchronousFetchOnInitialisation(false)
+                        .appName("test-sync-update")
+                        .unleashAPI("http://localhost:8080")
+                        .toggleBootstrapProvider(toggleBootstrapProvider)
+                        .build();
+        UnleashScheduledExecutor executor = mock(UnleashScheduledExecutor.class);
+        ToggleFetcher toggleFetcher = mock(ToggleFetcher.class);
+        ToggleBackupHandler toggleBackupHandler = mock(ToggleBackupHandler.class);
+        when(toggleBackupHandler.read())
+                .thenReturn(
+                        populatedToggleCollection(
+                                new FeatureToggle(
+                                        "toggleFeatureName1",
+                                        true,
+                                        Collections.singletonList(
+                                                new ActivationStrategy("custom", null))),
+                                new FeatureToggle(
+                                        "toggleFeatureName2",
+                                        true,
+                                        Collections.singletonList(
+                                                new ActivationStrategy("custom", null)))));
+        FeatureToggleRepository repo =
+                new FeatureToggleRepository(config, executor, toggleFetcher, toggleBackupHandler);
+        verify(toggleBootstrapProvider, times(0)).read();
+    }
+
+    private String fileToString(File f) throws IOException {
+        return new String(Files.readAllBytes(f.toPath()), StandardCharsets.UTF_8);
     }
 }